fix float for average ADC in datastore

+++ ELE_542_Projet/core/datastore.h
@@ -11,8 +11,8 @@
 
 typedef struct  
 {
-	uint32_t MoyenneADCDroit;
-	uint32_t MoyenneADCGauche;	
+	float MoyenneADCDroit;
+	float MoyenneADCGauche;	
 }datastore;
